+++ examples/api/src-tauri/tauri-plugin-sample/src/lib.rs
@@ -3,10 +3,18 @@ use tauri::{
   Runtime,
 };
 
-#[cfg(any(target_os = "android", target_os = "ios"))]
+use models::*;
+
+#[cfg(desktop)]
+mod desktop;
+#[cfg(mobile)]
 mod mobile;
-#[cfg(any(target_os = "android", target_os = "ios"))]
-pub use mobile::*;
+pub mod models;
+
+// Extensions to [`tauri::App`], [`tauri::AppHandle`] and [`tauri::Window`] to access the sample APIs.
+pub trait SampleExt<R: Runtime> {
+  fn ping(&self, payload: PingRequest) -> tauri::Result<Result<PingResponse, String>>;
+}
 
 pub fn init<R: Runtime>() -> TauriPlugin<R> {
   Builder::new("sample")
