+++ examples/api/src-tauri/tauri-plugin-sample/src/models.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize)]
+pub struct PingRequest {
+  pub value: Option<String>,
+}
+
+#[derive(Debug, Clone, Default, Deserialize)]
+pub struct PingResponse {
+  pub value: Option<String>,
+}
